requirements.txt
fastapi
uvicorn
starlette
mysql-connector-python
python-jose
python-multipart
opencv-contrib-python
pyzbar
pytesseract
pydantic
boto3
pillow
imutils
scipy
numpy
python-dateutil
cachetools
pytz
gunicorn
psutil
python-dotenv